+++ src/github.com/stellar/horizon/db/query_transaction_page.go
@@ -15,7 +15,7 @@ type TransactionPageQuery struct {
 
 func (q TransactionPageQuery) Select(ctx context.Context, dest interface{}) error {
 	sql := TransactionRecordSelect.
-		Limit(uint64(q.Limit))
+		Limit(q.Limit)
 
 	cursor, err := q.CursorInt64()
 	if err != nil {
